requirements.txt
fastapi[standard]
uvicorn[standard]
jinja2
aiortc
paho-mqtt
opencv-python
degirum
degirum_tools
degirum_cli